Finish rendering of tres format. This will likely need a number of modifications as I go into the future.

+++ src/Godot/Resources/StyleBoxFlat.php
@@ -13,19 +13,19 @@ class StyleBoxFlat extends StyleBox
     public float $contentMarginTop = 0;
     public float $contentMarginBottom = 0;
     public Color $bgColor;
-    public float $borderWidthLeft = 0;
-    public float $borderWidthTop = 0;
-    public float $borderWidthRight = 0;
-    public float $borderWidthBottom = 0;
+    public int $borderWidthLeft = 0;
+    public int $borderWidthTop = 0;
+    public int $borderWidthRight = 0;
+    public int $borderWidthBottom = 0;
     public Color $borderColor;
     public bool  $borderBlend = false;
-    public float $cornerRadiusTopLeft = 0;
-    public float $cornerRadiusTopRight = 0;
-    public float $cornerRadiusBottomRight = 0;
-    public float $cornerRadiusBottomLeft = 0;
+    public int $cornerRadiusTopLeft = 0;
+    public int $cornerRadiusTopRight = 0;
+    public int $cornerRadiusBottomRight = 0;
+    public int $cornerRadiusBottomLeft = 0;
 
     /**
-     * @param array<string,float|bool|Color|string|null> $options
+     * @param array<string,int|float|bool|Color|string|null> $options
      */
     public function __construct(array $options = [])
     {
@@ -74,28 +74,28 @@ class StyleBoxFlat extends StyleBox
 
     public function setBorderWidthLeft(float $value): static
     {
-        $this->borderWidthLeft = $value;
+        $this->borderWidthLeft = (int)$value;
 
         return $this;
     }
 
     public function setBorderWidthTop(float $value): static
     {
-        $this->borderWidthTop = $value;
+        $this->borderWidthTop = (int)$value;
 
         return $this;
     }
 
     public function setBorderWidthRight(float $value): static
     {
-        $this->borderWidthRight = $value;
+        $this->borderWidthRight = (int)$value;
 
         return $this;
     }
 
     public function setBorderWidthBottom(float $value): static
     {
-        $this->borderWidthBottom = $value;
+        $this->borderWidthBottom = (int)$value;
 
         return $this;
     }
@@ -119,28 +119,28 @@ class StyleBoxFlat extends StyleBox
 
     public function setCornerRadiusTopLeft(float $value): static
     {
-        $this->cornerRadiusTopLeft = $value;
+        $this->cornerRadiusTopLeft = (int)$value;
 
         return $this;
     }
 
     public function setCornerRadiusTopRight(float $value): static
     {
-        $this->cornerRadiusTopRight = $value;
+        $this->cornerRadiusTopRight = (int)$value;
 
         return $this;
     }
 
     public function setCornerRadiusBottomRight(float $value): static
     {
-        $this->cornerRadiusBottomRight = $value;
+        $this->cornerRadiusBottomRight = (int)$value;
 
         return $this;
     }
 
     public function setCornerRadiusBottomLeft(float $value): static
     {
-        $this->cornerRadiusBottomLeft = $value;
+        $this->cornerRadiusBottomLeft = (int)$value;
 
         return $this;
     }
